+++ src/index.js
@@ -4,10 +4,7 @@ import "./index.css";
 import App from "./App";
 import { makeServer } from "./server";
 import { BrowserRouter as Router } from "react-router-dom";
-import { CategoryProvider } from "./Context/category-provider"
-import { AuthProvider } from './Context/auth-provider'
-import { HistoryProvider } from './Context/history-provider'
-import { LikedVideoProvider, LikeedVideoProvider } from './Context/likevideo-provider'
+import { CategoryProvider, AuthProvider, HistoryProvider, LikedVideoProvider } from './Context/index'
 
 // Call make Server
 makeServer();
